backend/requirements/dev.txt
blue==0.9.1
flake8==4.0.1
ipython==8.11.0
pytest-django==4.5.2

-r local.txt